+++ ConstraidTests/FlushWithEdgesTests.swift
@@ -1,0 +1,111 @@
+import XCTest
+
+class FlushWithEdgesTests: XCTestCase {
+
+    func testFlushWithVerticalEdgesOf() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+        viewOne.flush(withVerticalEdgesOf: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+
+        let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
+        let constraintTwo = viewOne.constraints.last! as NSLayoutConstraint
+
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, NSLayoutAttribute.leading)
+        XCTAssertEqual(constraintOne.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, NSLayoutAttribute.leading)
+        XCTAssertEqual(constraintOne.constant, 10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, 500)
+
+        XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintTwo.firstAttribute, NSLayoutAttribute.trailing)
+        XCTAssertEqual(constraintTwo.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintTwo.secondAttribute, NSLayoutAttribute.trailing)
+        XCTAssertEqual(constraintTwo.constant, -10.0)
+        XCTAssertEqual(constraintTwo.multiplier, 2.0)
+        XCTAssertEqual(constraintTwo.priority, 500)
+    }
+
+    func testFlushWithHorizontalEdgesOf() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+        viewOne.flush(withHorizontalEdgesOf: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+
+        let constraintOne = viewOne.constraints.first! as NSLayoutConstraint
+        let constraintTwo = viewOne.constraints.last! as NSLayoutConstraint
+
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, NSLayoutAttribute.top)
+        XCTAssertEqual(constraintOne.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, NSLayoutAttribute.top)
+        XCTAssertEqual(constraintOne.constant, 10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, 500)
+
+        XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintTwo.firstAttribute, NSLayoutAttribute.bottom)
+        XCTAssertEqual(constraintTwo.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintTwo.secondAttribute, NSLayoutAttribute.bottom)
+        XCTAssertEqual(constraintTwo.constant, -10.0)
+        XCTAssertEqual(constraintTwo.multiplier, 2.0)
+        XCTAssertEqual(constraintTwo.priority, 500)
+    }
+
+    func testFlushWithEdgesOf() {
+        let viewOne = UIView()
+        let viewTwo = UIView()
+
+        viewOne.addSubview(viewTwo)
+        viewOne.flush(withEdgesOf: viewTwo, constant: 10.0, multiplier: 2.0, priority: 500)
+
+        let constraintOne = viewOne.constraints[0] as NSLayoutConstraint
+        let constraintTwo = viewOne.constraints[1] as NSLayoutConstraint
+        let constraintThree = viewOne.constraints[2] as NSLayoutConstraint
+        let constraintFour = viewOne.constraints[3] as NSLayoutConstraint
+
+        XCTAssertEqual(constraintOne.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintOne.firstAttribute, NSLayoutAttribute.top)
+        XCTAssertEqual(constraintOne.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintOne.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintOne.secondAttribute, NSLayoutAttribute.top)
+        XCTAssertEqual(constraintOne.constant, 10.0)
+        XCTAssertEqual(constraintOne.multiplier, 2.0)
+        XCTAssertEqual(constraintOne.priority, 500)
+
+        XCTAssertEqual(constraintTwo.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintTwo.firstAttribute, NSLayoutAttribute.bottom)
+        XCTAssertEqual(constraintTwo.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintTwo.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintTwo.secondAttribute, NSLayoutAttribute.bottom)
+        XCTAssertEqual(constraintTwo.constant, -10.0)
+        XCTAssertEqual(constraintTwo.multiplier, 2.0)
+        XCTAssertEqual(constraintTwo.priority, 500)
+
+        XCTAssertEqual(constraintThree.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintThree.firstAttribute, NSLayoutAttribute.leading)
+        XCTAssertEqual(constraintThree.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintThree.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintThree.secondAttribute, NSLayoutAttribute.leading)
+        XCTAssertEqual(constraintThree.constant, 10.0)
+        XCTAssertEqual(constraintThree.multiplier, 2.0)
+        XCTAssertEqual(constraintThree.priority, 500)
+
+        XCTAssertEqual(constraintFour.firstItem as! UIView, viewOne)
+        XCTAssertEqual(constraintFour.firstAttribute, NSLayoutAttribute.trailing)
+        XCTAssertEqual(constraintFour.relation, NSLayoutRelation.equal)
+        XCTAssertEqual(constraintFour.secondItem as! UIView, viewTwo)
+        XCTAssertEqual(constraintFour.secondAttribute, NSLayoutAttribute.trailing)
+        XCTAssertEqual(constraintFour.constant, -10.0)
+        XCTAssertEqual(constraintFour.multiplier, 2.0)
+        XCTAssertEqual(constraintFour.priority, 500)
+    }
+}